split off TestPeriodIndex class
